code cleanup, result_aggregators code moved into annotier
